add semiballistic version for missiles
missile launchers just won't fire at targets without direct targeting, though, so this is pretty useless

this would have to be a high-trajectory cannon? maybe? that instantly respawns as a missile (or something else ridiculous) to remain pure lua

this solution was fine to explore but a cpp option is just going to be better all around