enforce strict health values conditionally
Health less than 1 is not represented in the UI so is "not real" to the player. This is acceptable in the case of environmental set pieces meant to be destructible or fragile, however.

Allowing this awkward fractional exception is necessary for (e.g.) trees, which have a single hit point but take a constructor about two to eight seconds to reclaim (it depends).

I sacrificed some readability here. If you cannot figure the new flow, we first check for the feature's destruction, then for two cases: either the feature is damaged by a reclaim command or is "healed" ("repaired" is the wrong nomenclature, here) by a resurrect command.

Feature healing via resurrect is intended to slow the snowball effect (timing-constrained) but not punish recovery (resource-constrained). While the feature is still being healed, no resurrection progress can be made.