decrease type warnings from CMD
For example we get warnings for integer types with:
local CMD_STOP_INTEGER = CMD.STOP
local name = getCommandSection(gameCommands[getCommandCode(CMD_STOP_INTEGER)])

But we can receive a command id as a command parameter, e.g. from CMD_INSERT, and in other ways via the engine API functions (generally as either integer or number). We would prefer for these not to create warnings, even though `number` is an especially bad fit (0.5, pi, etc), because any id that does not map to a command is just an invalid command, not an exception.