add lookup and filter of command param counts
I'd avoided this because it serves only on real purpose (setting up the Printer and Fortifier) but maybe there is some justification. This would remove the entire hardcoded command config on the attached con turret scripts, replacing it with a set of commands linked to builder abilities (build, assist, capture, reclaim, repair, resurrect) and their non-area param sets (object or point only).