doc: these explicit checks might seem unusual
They are a side effect of requiring unique any/null sets. Testing a set to see if it conformed to these constraints would be ineffective for null sets and slow for the any set.

Wait commands are exceptional enough to warrant this check (it takes some time to realize that all other waits enqueue the base CMD_WAIT) but not to warrant the param type pointer to PRMTYPE_WAIT. We would need this only for processing on type pointers directly, but I want to _avoid_ this case because it presupposes the mapping between a prmType and the parameter index tables.

Lots of text to say: Helpful because it is clarifying, detracting because it is a bad example of how to employ this module.