specify wait command params
The family of wait commands has separate parameter sets between orders and enqueued commands. The enqueued command is always a CMD_WAIT with a WAITCODE to determine its sub-command and a unique wait key to handle ACKing, updating, and lifting the wait.

Poorly-formed waits can be enqueued but their handling will fail. The behavior is then outside the engine's handling and might be considered a game concern. The commands module is intended to allow registering new wait-type commands, so this is something we may support, though that is not especially likely.