add remove commands tricks
With CMD.BUILD part of the repo, we can start to eliminate (where appropriate) any implicit build order checks via `command < 0` and such. Since this is the commands module, the module made specifically for commands, command's module, yes that one, we can get away with this minimum-knowledge check, but in general we would want to write code that ignores it.