refactor: commit 2 of 2 replacing turret auto routine
This is the main, large refactor for the printer.

Most of the new code adds a pseudo-fight command update for idle construction turrets without relying on the attached def having certain properties. Actions are instead undertaken only if the attached con is determined to have that ability.

Note: This is not quite the same as a command -- something like CMD_IDLE_FIGHT would be a nice abstraction but isn't needed at the moment.

The rest introduces some structure for command introspection. `tryOrderDirection` checks that the turret's current target is within its actual build radius, acting as a double-check at this point, then gives us the heading to turn the turret to face (idk why the unit needs this -- are the animations not possible via the unit script itself?).

Finally, when the turret takes an independent action from the base, we can query the base to see if it would like to face toward the target position slightly. This can help with the very-slow open times of the printer's turret and with construction units in general, just by making them look like they're busy in the meantime.

The rest of this PR is focused on correcting disagreements between the base and turret unit to the degree of paranoia -- assuming a grand conspiracy of other gadgets and engine features against our implementation, how can we best keep the two units in sync with one another?