set a lower workertime and higher resurrectspeed
Time/speed change places in these unit def properties. They are eventually recalculated as inverses, properly, is why, so idk that I did this correctly.

This is dumb but in a fun way so I am going ahead with it. It is generally more robust practice to revise the unit def properties to give correct appearances when creating custom behavior, but this might be enough of a revision to resurrection that nothing can make it make sense to the unit AI or the skirmish or other AI.