use repairSpeed universally when repairing, even wrecks
This is honestly a slight perf drag that the gadget doesn't necessarily benefit from adding. Repair and rez speeds are very nearly merged together in the unit defs with not much impact here on gameplay. This is being right just to be right, and it's not even right.

So it is slightly daft, sure, but what if a modoption cranks repair up to x11 while leaving resurrect speed alone. Then we would look like real buffoons.