add ownership tug-o-war feature
The naive ownership implementation earlier would allow a single funneler to refill and reclaim from a single wreck to fund an ally tax-free. While that remains the case here -- the amount of metal sent to an allied owner is not taxed in this implementation -- I've added this ownership transfer feature.

Refilling a feature with metal just to reclaim it again will transfer ownership to the refiller over time. The result is that a unit's wreck can be used, really, only once. With extraordinary hustle, you could get multiple uses out of it, but I just do not see BAR players ever becoming so motivated.