add UnitInfo module for classifying units and using shared cache tables
The actual memory reduction is likely pitiful and the cache locality gains likewise but those are not the real reasons to do this. This aims to reduce repetitive code during basic config and to start providing basic recipes to gadgets and widgets that then can be kept maintained from a single place. Everything else is gravy.

An ideal might be for zero gadgets and widgets to reference the UnitDefs table directly (though that's going way too far). The *Defs tables are protected by Lua metatables (and maybe cpp stuff idk) and are just egregiously sluggish compared to a normal, fast-hashed lookup. Advice is to avoid these lookups like the plague but that can mean storing redundant data tables. This module should resolve any squeamishness about local caches.