ruin CEG and gfx flash brightness scaling
This restores the original damage total to the Starfall unit def, which makes the game think that we want flashes that are about 16x as large, bright, and distorted as they were previously. We now have to adjust everything related to the Starfall's gfx to restore any kind of order. This is also not a very good weapon def, just overall. There are problems to work out with how this should operate.