perf: cruise speceffect results cache
This is focused on the armsptk which fires a salvo of rockets, similar to the legmed, so it came to mind. Its performance drag seems to be high allocs and the (small) extra cost of the MOVE ops to push unused values into the "sink var".

So I guess my reasoning is that the sink var is not efficient, whereas this is very slightly more efficient. A table pool starts to look like a much better idea for reducing the allocations themselves, since these are what is slow/heavy.

If allocation really is this inefficient, then positionXYZ could be a float3, velocities a float4, and etc. with table reuse offsetting the overhead of table accesses. I'd have to test.