add colvol processing to collisionvolumes.lua
This is as relatively-nice as I could play with the existing unit_dynamic_collision_volume gadget, by moving all the preprocessing of colvols and mid/aim points to the configuration file, instead.

The costs incurred at UnitCreated are mostly in setting the colvols themselves, I assume. Still, getting all the processing done before creating the unit means that I can have a reference set of values for the unitDef. That means I can manipulate a colvol however I need to do so, then set it back when I'm done.

I wanted the late-loading of colvol data -- pAxis on unit colvols, piece data on piece colvols -- to be invisible or as close as possible to the gadget's use of that data. I think the metatables are fine but the global search for any unit with that unitDefID is stupid, whereas the gadget would know the answer to that trivially.

The unit defs with per-piece collision volumes are changed by this, actually, while everything else remains the same as before. Since they are added to `pieceCollisionVolume`, the code in UnitCreated does not reach the test for modelType to see if they are 3do or s3o model types. We have no 3do units, so only s3o is impacted. And the s3o rescaling only impacts air units with spherical colvols -- to whatever purpose. So the impact is mitigated by how stupid everything was in the first place, which I find beautiful.