rewrite unit_dynamic_collision_volume.lua
The functionality of this gadget was partially moved to collisionvolumes.lua and partially deprecated (unitmidpos, unitaimpos customparams are obsoleted) but it is in large part identical to before, otherwise.

Collision volumes are pre-processed and unit radius and height are no longer determined from either the model or the unitDef but some eldritch combination of the two, plus additional logic. We do not need to apply that logic live during gameplay, so all numbers and etc. are crunched upfront.

The classes/aliases for colvol types are bad and don't really validate anything in the gadget code. They are barely useful for configuration, at this point, too, since they include both the pre-processing config type defs and the post-processing config types used in this gadget (which, again, our poorly type checker cannot use to diagnose anything). Lua is not a good language for modern development for many reasons. This is one.

This also extends dynamic colvols slightly to a type of controller pattern where other gadgets can reserve permission to modify a collision volume. That is restricted to an anemic API surface for now since it is not yet used for anything.